Revert "Improve CTAs (#908)"
This reverts commit 8df5bc3af9f8e3ae1a6fb43b7f73a95a019d71d3.
