--- src/components/BlogFooter/style.scss
@@ -1,71 +1,0 @@
-.blog-footer-cta-wrapper {
-    margin-top: -20px;
-
-    .blog-footer-divider {
-        max-width: 90vw !important;
-        width: 800px !important;
-        margin-left: auto !important;
-        margin-right: auto !important;
-    }
-
-    .blog-footer-img {
-        transform: rotateY(180deg);
-        position: absolute;
-        right: 0;
-        margin-top: 10px;
-    }
-
-    .blog-footer-description {
-        max-width: 400px;
-        font-size: 14px;
-        text-align: justify;
-        text-justify: inter-word;
-    }
-
-    .blog-footer-row {
-        max-width: 800px;
-        margin: auto;
-    }
-
-    @media screen and (max-width: 1076px) {
-        .blog-footer-description {
-            font-size: 14px !important;
-            margin-right: 30px;
-        }
-        .blog-footer-img {
-            right: 50px !important;
-        }
-    }
-
-    @media screen and (max-width: 560px) {
-        .blog-footer-img {
-            margin-top: 40px;
-        }
-
-        .blog-footer-img {
-            right: 0 !important;
-        }
-
-        h4 {
-            font-size: 18px !important;
-        }
-
-        .ant-col-14 {
-            width: 65%;
-        }
-
-        .ant-col-8 {
-            width: 35%;
-        }
-    }
-
-    @media screen and (max-width: 350px) {
-        .blog-footer-btn {
-            font-size: 14px !important;
-            padding: 2px !important;
-        }
-        .blog-footer-description {
-            font-size: 12px !important;
-        }
-    }
-}
