+++ src/components/Layout/Layout.scss
@@ -1,6 +1,3 @@
-@import '../../vars';
-
-// Some comment
 /* latin */
 @font-face {
     font-family: 'Lato';
@@ -670,10 +667,14 @@ header {
     align-items: center;
 }
 
-.header-key {
+.headerKey {
     margin-left: 20px;
     float: 'right';
     margin-bottom: 'calc(1.45rem / 2)';
+}
+
+.headerKey a {
+    color: #000000;
 }
 
 .mobileHeader {
@@ -795,7 +796,7 @@ aside.sidebar {
     color: #1d4aff !important;
 }
 
-.keySelectedLink:hover {
+.ant-menu-item-active > a {
     color: #1d4aff !important;
 }
 
@@ -823,7 +824,7 @@ li.ant-menu-submenu-selected > div {
 }
 
 .ant-menu-item-selected a {
-    color: #1d4aff;
+    color: #1d4aff !important;
 }
 
 aside.sidebarRight {
@@ -1204,6 +1205,20 @@ table th {
     color: #ffffff !important;
 }
 
+.login-btn.blogPage:hover,
+.headerItem.blogPage:hover,
+.headerItem:hover {
+    color: #1d4aff !important;
+}
+
+.login-btn {
+    color: #1d4aff !important;
+}
+
+.login-btn:hover {
+    color: #577bfb !important;
+}
+
 .centered {
     display: block;
     text-align: center;
@@ -1330,10 +1345,6 @@ li {
 .announcement-banner:hover {
     background-color: #fab938;
     cursor: pointer;
-}
-
-#navbar-responsive-wrapper li.responsive-menu-item {
-    border-bottom: 1px solid #e8e8e8 !important;
 }
 
 @media screen and (min-width: 1076px) {
@@ -1481,53 +1492,6 @@ for all other rows the portrait image wraps above bio*/
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
-}
-
-.main-nav-cta {
-    border: 1px solid $border;
-    border-radius: 2px;
-    padding: 8px 16px;
-    color: #fff;
-    transition: background-color 300ms ease-out;
-    box-shadow: 0 2px 0 rgba(0, 0, 0, 0.045);
-}
-
-.main-nav-cta-wrapper:hover {
-    opacity: inherit !important;
-}
-
-.main-nav-cta-dark-bg {
-    background-color: white;
-    color: $text_default;
-}
-
-.main-nav-cta-light-bg {
-    background-color: $primary;
-    border-color: $primary;
-}
-
-.main-nav-cta-wrapper:hover {
-    .main-nav-cta-light-bg {
-        background-color: darken($primary, 5%);
-    }
-
-    .main-nav-cta-dark-bg {
-        background-color: darken(white, 20%);
-    }
-}
-
-.header-key:hover {
-    opacity: 0.5;
-}
-
-.header-key:visited {
-    color: inherit !important;
-}
-
-.header-key-active span {
-    padding-bottom: 4px;
-    border-bottom: solid 2px $brand_blue;
-    color: $brand_blue;
 }
 
 /*split to one column on smaller screens*/
