+++ src/components/Menu/index.js
@@ -1,0 +1,3 @@
+import Menu from './Menu'
+
+export default Menu
