+++ src/logic/layoutLogic.js
@@ -15,7 +15,6 @@ export const layoutLogic = kea({
         onSidebarContentExpanded: (expandedKeys) => ({ expandedKeys }),
         onSidebarContentSelected: (selectedKey) => ({ selectedKey }),
         onChangeMenuState: (nItem) => ({ nItem }),
-        setIsGetStartedModalOpen: (open) => ({ open }),
     },
 
     reducers: {
@@ -103,12 +102,6 @@ export const layoutLogic = kea({
             null,
             {
                 onChangeMenuState: (state, { nItem }) => nItem,
-            },
-        ],
-        isGetStartedModalOpen: [
-            false,
-            {
-                setIsGetStartedModalOpen: (_, { open }) => open,
             },
         ],
     },
