+++ src/pages/styles/index.scss
@@ -1429,3 +1429,158 @@
         }
     }
 }
+
+/* MODAL */
+.modalOverlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(255, 255, 255, 0.3);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 9000;
+}
+
+.modalContent {
+    position: relative;
+    background: #f4f4f4;
+    overflow: auto;
+    -webkit-overflow-scrolling: touch;
+    padding: 20px;
+    width: 920px;
+    height: 496px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-evenly;
+    z-index: 9001;
+    box-shadow: 0px 26px 26px rgba(0, 0, 0, 0.25);
+    border-radius: 24px;
+}
+
+.modalCardsWrapper {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    width: 100%;
+}
+
+.modalCard {
+    position: relative;
+    width: 280px;
+    height: 340px;
+    background: #ffffff;
+    box-shadow: 8px 8px 12px rgba(0, 0, 0, 0.25);
+    border-radius: 20px;
+    padding: 32px;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
+
+.modalCard:hover {
+    border: 4px solid #1d4aff;
+    padding: 28px;
+}
+
+.modalCardHeader {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: flex-start;
+}
+
+.modalCard h2 {
+    font-size: 32px !important;
+    line-height: 40px;
+}
+
+.modalCard h4 {
+    font-family: Arial, Helvetica, sans-serif;
+    color: #f96132;
+    font-size: 14px;
+    line-height: 24px;
+    font-weight: bold;
+}
+
+.modalCard p {
+    font-size: 14px;
+    line-height: 24px;
+}
+
+button.modalClose {
+    position: absolute;
+    right: 16px;
+    top: 16px;
+    background-color: #ababab;
+    border-color: #ababab;
+    color: #fff;
+    transition: all 0.3s;
+    height: 24px;
+    width: 24px;
+    font-size: 14px;
+}
+
+button.modalClose:hover {
+    background-color: #fff;
+    border-color: #1d4aff;
+    color: #1d4aff;
+}
+
+@media screen and (max-width: 834px) {
+    .modalContent {
+        height: 90vh;
+        width: 90vw;
+    }
+    .modalCardsWrapper {
+        height: 700px;
+        flex-direction: column;
+        justify-content: space-between;
+        align-items: center;
+    }
+    .modalCard {
+        width: 85vw;
+        height: 23vh;
+        padding: 20px 35px;
+    }
+    .modalCardHeader {
+        width: 218px;
+        height: 64px;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        align-items: center;
+        margin-bottom: 12px;
+    }
+
+    .modalCardHeader img {
+        width: 40px !important;
+    }
+
+    .modalCardHeader img,
+    .modalCardHeader h2 {
+        margin: 0 !important;
+    }
+    .modalContent h2 {
+        font-size: 18px !important;
+        line-height: 20px !important;
+    }
+    .modalContent p {
+        display: none;
+    }
+
+    .modalContent h4 {
+        font-size: 14px;
+        line-height: 18px;
+        font-weight: 200;
+        color: #000000;
+    }
+    .modalContent > h2 {
+        text-align: center;
+        margin-top: 10px;
+        width: 207px;
+    }
+}
